+++ libmtx/vector/omp.h
@@ -97,7 +97,7 @@ int mtxvector_omp_alloc(
     struct mtxvector_omp * x,
     enum mtxfield field,
     enum mtxprecision precision,
-    int size,
+    int64_t size,
     int num_threads);
 
 /**
@@ -106,7 +106,7 @@ int mtxvector_omp_alloc(
  */
 int mtxvector_omp_init_real_single(
     struct mtxvector_omp * x,
-    int size,
+    int64_t size,
     const float * data,
     int num_threads);
 
@@ -116,7 +116,7 @@ int mtxvector_omp_init_real_single(
  */
 int mtxvector_omp_init_real_double(
     struct mtxvector_omp * x,
-    int size,
+    int64_t size,
     const double * data,
     int num_threads);
 
@@ -126,7 +126,7 @@ int mtxvector_omp_init_real_double(
  */
 int mtxvector_omp_init_complex_single(
     struct mtxvector_omp * x,
-    int size,
+    int64_t size,
     const float (* data)[2],
     int num_threads);
 
@@ -136,7 +136,7 @@ int mtxvector_omp_init_complex_single(
  */
 int mtxvector_omp_init_complex_double(
     struct mtxvector_omp * x,
-    int size,
+    int64_t size,
     const double (* data)[2],
     int num_threads);
 
@@ -146,7 +146,7 @@ int mtxvector_omp_init_complex_double(
  */
 int mtxvector_omp_init_integer_single(
     struct mtxvector_omp * x,
-    int size,
+    int64_t size,
     const int32_t * data,
     int num_threads);
 
@@ -156,7 +156,7 @@ int mtxvector_omp_init_integer_single(
  */
 int mtxvector_omp_init_integer_double(
     struct mtxvector_omp * x,
-    int size,
+    int64_t size,
     const int64_t * data,
     int num_threads);
